Resolve some CodeClimate issues
