+++ src/utils.js
@@ -487,6 +487,37 @@ function correctSelectorIdentifier(value) {
   return enquoteHelper.join('');
 }
 
+/* FBX Exporter stuff */
+/**
+ * Extension to standard Float32Arrays.
+ * This is the fastest way to concat TypedArrays in js (by observations).
+ * @param{Float32Array} first  - destination array
+ * @param{Float32Array} second - source array
+ * @returns{Float32Array} resulting concatenated array
+ */
+function Float32Concat(first, second) {
+  const firstLength = first.length;
+  const result = new Float32Array(firstLength + second.length);
+  result.set(first);
+  result.set(second, firstLength);
+  return result;
+}
+
+/**
+ * Extension to standard Int32Arrays.
+ * This is the fastest way to concat TypedArrays in js (by observations).
+ * @param{Int32Array} first  - destination array
+ * @param{Int32Array} second - source array
+ * @returns{Int32Array} resulting concatenated array
+ */
+function Int32Concat(first, second) {
+  const firstLength = first.length;
+  const result = new Int32Array(firstLength + second.length);
+  result.set(first);
+  result.set(second, firstLength);
+  return result;
+}
+
 //----------------------------------------------------------------------------
 // Exports
 
@@ -523,4 +554,6 @@ export default {
   getFileExtension,
   splitFileName,
   download,
+  Float32Concat,
+  Int32Concat,
 };
